Added Integration Test Case.
Whenever we use Mockery, Patchwork, or Brain Monkey, we need to ensure
the packages are properly set up and torn down.  These tasks are handled
in the Integration Test Case.

Moved the `format_the_html()` method the Test Case for use in
other tests besides just the Fields API.
