+++ tests/phpunit/integration/api/fields/includes/class-fields-test-case.php
@@ -9,14 +9,14 @@
 
 namespace Beans\Framework\Tests\Integration\API\Fields\Includes;
 
-use WP_UnitTestCase;
+use Beans\Framework\Tests\Integration\Test_Case;
 
 /**
  * Abstract Class Fields_Test_Case
  *
  * @package Beans\Framework\Tests\Integration\API\Fields\Includes
  */
-abstract class Fields_Test_Case extends WP_UnitTestCase {
+abstract class Fields_Test_Case extends Test_Case {
 
 	/**
 	 * An array of test data.
@@ -136,33 +136,5 @@ abstract class Fields_Test_Case extends WP_UnitTestCase {
 		}
 
 		return $field;
-	}
-
-	/**
-	 * Format the HTML by stripping out the whitespace between the HTML tags and then putting each tag on a separate
-	 * line.
-	 *
-	 * Why? We can then compare the actual vs. expected HTML patterns without worrying about tabs, new lines, and extra
-	 * spaces.
-	 *
-	 * @since 1.5.0
-	 *
-	 * @param string $html HTML to strip.
-	 *
-	 * @return string
-	 */
-	protected function format_the_html( $html ) {
-		$html = trim( $html );
-
-		// Strip whitespace between the tags.
-		$html = preg_replace( '/(\>)\s*(\<)/m', '$1$2', $html );
-
-		// Strip whitespace at the end of a tag.
-		$html = preg_replace( '/(\>)\s*/m', '$1$2', $html );
-
-		// Strip whitespace at the start of a tag.
-		$html = preg_replace( '/\s*(\<)/m', '$1$2', $html );
-
-		return str_replace( '>', ">\n", $html );
 	}
 }
