+++ tests/phpunit/integration/bootstrap.php
@@ -68,3 +68,6 @@ tests_add_filter( 'setup_theme', function() {
 
 // Start up the WP testing environment.
 require_once $beans_tests_dir . '/includes/bootstrap.php';
+
+// Load the Integration Test Case.
+require_once BEANS_TESTS_DIR . '/class-test-case.php';
